Create add transaction activity

+++ app/src/main/java/com/togglecorp/paiso/Debt.java
@@ -1,14 +1,14 @@
 package com.togglecorp.paiso;
 
 public class Debt {
-    public String by;
-    public String to;
-    public int amount;
+    public String by;       // Is either @me or contact-id
+    public String to;       // Is either @me or contact-id
+    public float amount;
     public String unit = "Rs.";
 
     public Debt() {}
 
-    public Debt(String by, String to, int amount, String unit) {
+    public Debt(String by, String to, float amount, String unit) {
         this.by = by;
         this.to = to;
         this.amount = amount;
